requirements.txt
attrs==18.1 --hash=sha256:4b90b09eeeb9b88c35bc642cbac057e45a5fd85367b985bd2809c62b7b939265
websockets==5.0 --hash=sha256:4ea90a598b546233a4c61bef98db4171ac30f1302826729af30001ac1f47109
